+++ android/src/main/kotlin/com/fluttercandies/flutter_qweather/MethodConstants.java
@@ -3,20 +3,33 @@ package com.fluttercandies.flutter_qweather;
 /// 方法常量
 public class MethodConstants {
     /// 获取平台版本
-    public static final String GetPlatformVersion = "getPlatformVersion";
+    public static final String GetPlatformVersion = "GetPlatformVersion";
     /// 设置 Debug
-    public static final String SetDebug = "setDebug";
+    public static final String SetDebug = "SetDebug";
     /// 初始化
-    public static final String Init = "init";
+    public static final String Init = "Init";
+
+    // 城市信息查询
+    public static final String GeoCityLookup = "GeoCityLookup";
+
+    // 热门城市查询
+    public static final String GetGeoTopCity = "GetGeoTopCity";
+
+    // POI信息搜索
+    public static final String GeoPoiLookup = "GeoPoiLookup";
+
+    // POI范围搜索
+    public static final String GeoPoiRangeLookup = "GeoPoiRangeLookup";
+
     /// 获取实况天气数据
-    public static final String GetWeatherNow = "getWeatherNow";
+    public static final String GetWeatherNow = "GetWeatherNow";
 
     // 获取逐天预报
-    public static final String GetWeatherDaily = "getWeatherDaily";
+    public static final String GetWeatherDaily = "GetWeatherDaily";
 
     // 获取逐时预报
-    public static final String GetWeatherHourly = "getWeatherHourly";
+    public static final String GetWeatherHourly = "GetWeatherHourly";
 
     // 获取中国地区未来2小时内每5分钟降水
-    public static final String GetWeatherMinuteLy = "getWeatherMinuteLy";
+    public static final String GetWeatherMinuteLy = "GetWeatherMinuteLy";
 }