+++ ios/Classes/Api/ApiWeather.m
@@ -1,6 +1,6 @@
 //
 //  ApiWeather.m
-//  qweather
+//  flutter_qweather
 //
 //  Created by CyJay on 2021/4/19.
 //
@@ -14,8 +14,7 @@
 + (void) getWeatherNow:(id _Nullable)param result:(FlutterResult)result{
     QWeatherConfigInstance.location = param;
     [QWeatherConfigInstance weatherWithInquireType:INQUIRE_TYPE_WEATHER_NOW
-      WithSuccess:^(id responseObject) {
-        WeatherBaseClass *rep = responseObject;
+      WithSuccess:^(WeatherBaseClass *rep) {
         [DebugPrint print:[@"getWeatherNow WithSuccess: " stringByAppendingString:rep.description]];
         if (![rep.code isEqualToString:@"200"]){
             result(NULL);
@@ -53,8 +52,7 @@
         inquireType = INQUIRE_TYPE_WEATHER_15D;
     }
     [QWeatherConfigInstance weatherWithInquireType:inquireType
-      WithSuccess:^(id responseObject) {
-        WeatherBaseClass *rep = responseObject;
+      WithSuccess:^(WeatherBaseClass *rep) {
         [DebugPrint print:[@"getWeatherDaily WithSuccess: " stringByAppendingString:rep.description]];
         if (![rep.code isEqualToString:@"200"]){
             result(NULL);
@@ -64,11 +62,11 @@
         for (Daily *one in rep.daily) {
             NSDictionary *oneDaily = @{
                 @"cloud": one.cloud, @"fxDate": one.fxDate, @"humidity":one.humidity,@"iconDay":one.iconDay, @"iconNight": one.iconNight,
-                @"moonPhase": one.moonPhase, @"moonrise": one.moonrise, @"moonset": one.moonset, @"moonrise": one.moonrise,
-                @"precip": one.precip, @"pressure": one.pressure, @"sunrise": one.sunrise, @"sunset": one.sunset, @"tempMax": one.tempMax,
-                @"tempMin": one.tempMin, @"textDay": one.textDay, @"textNight": one.textNight, @"uvIndex": one.uvIndex, @"vis": one.vis,
-                @"wind360Day": one.wind360Day, @"wind360Night": one.wind360Night, @"windDirDay": one.windDirDay, @"windDirNight": one.WindDirNight,
-                @"windScaleDay": one.windScaleDay, @"windScaleNight": one.windScaleNight,@"windSpeedDay": one.windSpeedDay, @"windSpeedNight": one.windSpeedNight
+                @"moonPhase": one.moonPhase, @"moonset": one.moonset, @"moonrise": one.moonrise, @"precip": one.precip, @"pressure": one.pressure,
+                @"sunrise": one.sunrise, @"sunset": one.sunset, @"tempMax": one.tempMax, @"tempMin": one.tempMin, @"textDay": one.textDay,
+                @"textNight": one.textNight, @"uvIndex": one.uvIndex, @"vis": one.vis, @"wind360Day": one.wind360Day, @"wind360Night": one.wind360Night,
+                @"windDirDay": one.windDirDay, @"windDirNight": one.WindDirNight, @"windScaleDay": one.windScaleDay, @"windScaleNight": one.windScaleNight,
+                @"windSpeedDay": one.windSpeedDay, @"windSpeedNight": one.windSpeedNight
             };
             [daily addObject:oneDaily];
         }
@@ -97,8 +95,7 @@
         inquireType = INQUIRE_TYPE_WEATHER_168H;
     }
     [QWeatherConfigInstance weatherWithInquireType:inquireType
-      WithSuccess:^(id responseObject) {
-        WeatherBaseClass *rep = responseObject;
+      WithSuccess:^(WeatherBaseClass *rep) {
         [DebugPrint print:[@"getWeatherHourly WithSuccess: " stringByAppendingString:rep.description]];
         if (![rep.code isEqualToString:@"200"]){
             result(NULL);
@@ -129,8 +126,7 @@
 + (void) getWeatherMinuteLy:(id _Nullable)param result:(FlutterResult)result{
     QWeatherConfigInstance.location = param;
     [QWeatherConfigInstance weatherWithInquireType:INQUIRE_TYPE_WEATHER_MINUTELY
-      WithSuccess:^(id responseObject) {
-        WeatherMinutelyBaseClass *rep = responseObject;
+      WithSuccess:^(WeatherMinutelyBaseClass *rep) {
             [DebugPrint print:[@"getWeatherMinuteLy WithSuccess: " stringByAppendingString:rep.description]];
         if (![rep.code isEqualToString:@"200"]){
             result(NULL);