+++ ios/Classes/MethodConstants.h
@@ -5,16 +5,24 @@
 //  Created by CyJay on 2021/4/18.
 //
 // 获取平台版本
-#define GetPlatformVersion @"getPlatformVersion"
+#define GetPlatformVersion @"GetPlatformVersion"
 // 设置 Debug
-#define SetDebug @"setDebug"
+#define SetDebug @"SetDebug"
 // 初始化
-#define Init @"init"
+#define Init @"Init"
+// 城市信息查询
+#define GeoCityLookup @"GeoCityLookup"
+// 热门城市查询
+#define GetGeoTopCity @"GetGeoTopCity"
+// POI信息搜索
+#define GeoPoiLookup @"GeoPoiLookup"
+// POI范围搜索
+#define GeoPoiRangeLookup @"GeoPoiRangeLookup"
 // 获取实况天气数据
-#define GetWeatherNow @"getWeatherNow"
+#define GetWeatherNow @"GetWeatherNow"
 // 获取逐天预报
-#define GetWeatherDaily @"getWeatherDaily"
+#define GetWeatherDaily @"GgetWeatherDaily"
 // 获取逐时预报
-#define GetWeatherHourly @"getWeatherHourly"
+#define GetWeatherHourly @"GetWeatherHourly"
 // 获取中国地区未来2小时内每5分钟降水
-#define GetWeatherMinuteLy @"getWeatherMinuteLy"
+#define GetWeatherMinuteLy @"GetWeatherMinuteLy"