+++ lib/constants.dart
@@ -1,25 +1,37 @@
 /// 方法常量
 class MethodConstants {
   // 获取平台版本
-  static const String GetPlatformVersion = 'getPlatformVersion';
+  static const String GetPlatformVersion = 'GetPlatformVersion';
 
   // 初始化
-  static const String Init = 'init';
+  static const String Init = 'Init';
 
   // 设置 Debug
-  static const String SetDebug = 'setDebug';
+  static const String SetDebug = 'SetDebug';
+
+  // 城市信息查询
+  static const String GeoCityLookup = 'GeoCityLookup';
+
+  // 热门城市查询
+  static const String GetGeoTopCity = 'GetGeoTopCity';
+
+  // 热门城市查询
+  static const String GeoPoiLookup = 'GeoPoiLookup';
+
+  // POI范围搜索
+  static const String GeoPoiRangeLookup = 'GeoPoiRangeLookup';
 
   // 获取实况天气数据
-  static const String GetWeatherNow = 'getWeatherNow';
+  static const String GetWeatherNow = 'GetWeatherNow';
 
   // 获取逐天预报
-  static const String GetWeatherDaily = 'getWeatherDaily';
+  static const String GetWeatherDaily = 'GetWeatherDaily';
 
   // 获取逐时预报
-  static const String GetWeatherHourly = 'getWeatherHourly';
+  static const String GetWeatherHourly = 'GetWeatherHourly';
 
   // 获取中国地区未来2小时内每5分钟降水
-  static const String GetWeatherMinuteLy = 'getWeatherMinuteLy';
+  static const String GetWeatherMinuteLy = 'GetWeatherMinuteLy';
 }
 
 /// 逐天可选枚举
@@ -35,4 +47,14 @@ enum WeatherHourlyForecast {
   WeatherForecast24Hour, // 未来 24 小时预告
   WeatherForecast72Hour, // 未来 72 小时预告
   WeatherForecast168Hour // 未来 168 小时预告
+}
+
+///枚举类型转string
+String enumToString(o) => o.toString().split('.').last;
+
+/// POI类型
+enum PoiType {
+  scenic, // 景点
+  CSTA, // 潮流站点
+  TSTA, // 潮汐站点
 }