+++ CHANGELOG.md
@@ -1,8 +1,18 @@
+## 0.0.3
+
+- realized geo search.
+- 实现 Geo 搜索
+
 ## 0.0.2
-  - 实现 逐天天气查询
-  - 实现 逐时天气查询
-  - 实现 中国地区未来2小时内每5分钟降水查询
+
+- realized daily weather query.
+- realized hourly weather query.
+- realized Precipitation query every 5 minutes in the next 2 hours in China.
+- 实现 逐天天气查询
+- 实现 逐时天气查询
+- 实现 中国地区未来2小时内每5分钟降水查询
 
 ## 0.0.1
 
-  - 实现 实时天气查询。
+- realized now weather query.
+- 实现 实时天气查询。